refactor: canister suffix
