+++ packages/nns/src/genesis_token.canister.spec.ts
@@ -1,7 +1,7 @@
 import { ActorSubclass } from "@dfinity/agent";
 import { mock } from "jest-mock-extended";
 import type { _SERVICE as GenesisTokenService } from "../candid/genesis_token";
-import { GenesisTokenCanister } from "./genesis_token";
+import { GenesisTokenCanister } from "./genesis_token.canister";
 
 describe("GenesisTokenCanister", () => {
   it("claimNeurons returns a list of neuron ids", async () => {
