+++ packages/nns/src/index.ts
@@ -2,10 +2,10 @@ export type { DeployedSns } from "../candid/sns_wasm";
 export { AccountIdentifier, SubAccount } from "./account_identifier";
 export * from "./errors/governance.errors";
 export * from "./errors/ledger.errors";
-export { GenesisTokenCanister } from "./genesis_token";
-export { GovernanceCanister } from "./governance";
+export { GenesisTokenCanister } from "./genesis_token.canister";
+export { GovernanceCanister } from "./governance.canister";
 export { ICP } from "./icp";
-export { LedgerCanister } from "./ledger";
+export { LedgerCanister } from "./ledger.canister";
 export * from "./types/common";
 export * from "./types/governance.options";
 export * from "./types/governance_converters";
