+++ packages/nns/src/sns_wasm.canister.spec.ts
@@ -2,7 +2,7 @@ import { ActorSubclass } from "@dfinity/agent";
 import { mock } from "jest-mock-extended";
 import type { _SERVICE as SnsWasmService } from "../candid/sns_wasm";
 import { snsMock } from "./mocks/sns_wasm.mock";
-import { SnsWasmCanister } from "./sns_wasm";
+import { SnsWasmCanister } from "./sns_wasm.canister";
 
 describe("Sns-wasm", () => {
   it("should return the list of sns", async () => {
